requirements.txt
flask==3.0.2
requests==2.31.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
flask-socketio==5.3.6
eventlet==0.35.2
websocket-client==1.7.0
werkzeug==3.0.1
python-socketio==5.11.2